This project is a Object for a 2D Array Linked List, This list is single linked LinkedList.
This class creates a 2DLinked List. This Linked List will have references to the head, references to the node that is first in the last column, and reference to the node that first for the last row.

The 2D Array class has methods that can, add rows and columns at either first or last sections of the list. and can also add rows or columns at specific indexes. When adding rows and columns the amount of values need to be the same length as the rows or columns. The reason for this is to keep the LinkedList in a even format.

There are also delete methods for rows and columns for either the first or last selection or a specific index.

There is also a updated toString method that will print out the LinkedList correctly.